Make source panel functional and refactor
