+++ src/components/SourcePanel/index.ts
@@ -1,0 +1,1 @@
+export {default as SourcePanel} from './SourcePanel';
